Revamped the SMTP client a bit and changed the server for the SMTP example.
The SMTP client now prints the command that was issued whenever the server sends an error so that problems can be better debugged. It now also obeys the settings.connectionType field and settings.useTLS was removed in turn.

The SMTP example uses smtp.example.com now so that it already fails looking up the host and does not try to send a mail using some actually existing server. See also issue #126.