Refactored the Diet template compiler.
Building of the final D code is now abstracted in OutputContext so that changes can be made centrally and the rest of the code is more readable (e.g. no double escaped string literals).

Also string concatenations in the generated code are avoided now to safe on memory allocations. There are still some calls to _toString and htmlEscape that need to be eliminated so that there are no allocations left.