Revert the path optimization.
Was corrupting results as soon as a Path object was moved. Unfortunately, because of the support for sub paths and some other features, this kind of optimization is theoretically not possible. An alternative should be to either avoid use of Url/Path in the HTTP server/client or to extend them to use a custom allocator.