Implement automatic help screen functionality, adjust code style and move privilege lowering code to vibe.core.core.
Privilege lowering was broken as it happened before any calls to listenTCP() or
similar functions. It is now executed explicitly using lowerPrivileges(). Also
processCommandLineArgs() is now fixed to at least perform in a reasonably
similar way to how it used to do so that compatibility is not silently broken.

See also #211.