Fix connection delays in the libevent2 driver. Fixes #338.
The 4096 low write watermark caused write operations to stall when no more than 4096 bytes were sent. The current behavior with a zero watermark now has the same behavior as the libev and the win32 driver (i.e. writing all data to the socket before returning from write()). It still needs to be determined if this is relevant for benchmark results and using a higher low watermark should be fixed instead.

Additionally, the more error prone "writeFinished" field has been removed and replaced by an explicit check of the output buffer size.