Set a maximum read buffer for Libevent2TCPConnection to avoid huge memory consumption and other side effects.
See http://forum.rejectedsoftware.com/groups/rejectedsoftware.vibed/thread/7927/ for "other" side effects. Note that this issue has another root cause, which is that .empty should not be true unless the read buffer has been fully emptied. Currently it is true as soon as the connection is closed by the remote.