Add a first version of an ORM with a MongoDB driver for the start.
The current proof-of-concept functionality is limited to insertion and simple queries.

The main goals, which are not achieved by the existing solutions for D, are the avoidance of memory allocations for optimum performance and the use of statically checked query expressions, at least for the most common queries. Avoiding memory allocations is achieved by making use of templates and related meta programming tools and by using direct serialization of D types instead of storing a database specific data format in memory. Note, however, that this is still dependent on the database driver and the current MongoDB driver does perform some allocations.