Fix `RedisClient` instantiation for authentication
This removes the "INFO" command from the default `RedisClient`
constructor, since its only there to support the `redisVersion` getter.
This fixes creating clients for databases that require authentication
for all commands. It moves the version fetching logic into the
`redisVersion` getter, and caches its result in the `m_version`
property. I think this is the cleanest solution to this issue.