Fix Websocket pinging.
When vibe.d received a ping frame, it set the frame's opcode to pong and sent the pong message, it broke out of the switch and the loop as well (since the opcode is no longer ping), which left the m_currentFrame in a zeroed out state so Websocket receive complained about unexpected continuation frame and closed the connection.

Heartbeat made the connection go away - sweet irony.