Avoid access of possible dangling driver references. Fixes #978.
This is a two-step fix: destroy() is used instead of delete for the event driver and the driver core, so that there can not be any dangling references left. Additionally, the Timer struct now only tries to release the timer counter when the library has not yet been shut down. This avoids triggering an assertion, which in turn results in the InvalidMemoryOperationError in #978.