REST: Add unittest to reject ref / out body and query parameters
Those don't make sense, as it's impossible to return via the query,
and it doesn't make sense to return via the body in such a complicated
way when we can do it via the return value of the function.
Header parameters on the other hand could be legitimately marked as
ref or out. However the feature is not yet implemented and will come
in a future patch.