REST/WEB: Ensure that we have a non-empty URL
Previously, having a function named with one of the prefixes
(e.g: `get`, `post`, `query`) would result in a different
behavior in the web and rest interface. The REST interface
will generate faulty code that could crash at some point,
while the web generator will treat that as a '/'.

Instead of special-casing this, this commit makes it an error.
This error will be triggered when `registerXXXInterface` is
called, thus it is a CT error, but it was not possible to
make it a deprecation since we're dealing with CTFE.