Fix shutdown sequence.
The main thread used to remove itself from the thread list prior to waiting for the worker threads to shut down. This could cause the worker threads to think that they are in fact the main thread, because now they had the index 0. This change delays removing the main thread until all other threads have shut down.