Let conversion errors of web path parameters result in a 404 error.
It used to result in a 500 internal server error, but a URL that fails conversion for certain URL path parts obviously doesn't correspond to an existing resource, so a 404 is preferable. (a bad request error is also preferable to 500, but that's a different story)

This also has the implication that fall-through routes now work properly:

@path("/foo/:num") void foo(int _num) {}
@path("/foo/bar") void bar(){}

This will now properly fall-trough the foo() route if /foo/bar is requested, where previously the two routes would have to be reversed for the bar() route to be accessible at all.