Implement Diet template caching.
This allows downstream projects to define -version=VibeCacheDietTemplates to let the compiled Diet template (D code) be cached in the "views" folder, so that it gets picked up on the next compiler run instead of recompiling the template again. This can vastly improve compile time and memory use during edit-compile-run cycles.

Note that the cached source code is written at application initialization time. The program must be run from the root of the package directory, so that the cached files get written into the right "views" folder.