websocket: Simplify server / client difference
Before, the code used to have a flag for server code, and pass around
a SystemRNG as source of entropy.
This was a bit weird as their two parameters where not explicitly tied
together.
With this change, we assume that the presence of a source of entropy
means it's a client socket, and a `null` value means its a server.