websocket: Correct length for Sec-WebSocket-Key during protocol upgrade
The RFC specify in section 4.1 (Opening Handshake, Client Requirements) that:
"The value of this header field MUST be a nonce consisting of a randomly
selected 16-byte value that has been base64-encoded"

However `randomUUID.toString` would return a 36 chars string.
As a result, the base64 encoding was 48 chars long instead of 24.

Part of rejectedsoftware/vibe.d#1505