Don't force immutability for strings
When taking a string argument for validation, 'const(char)[]' should be used instead of 'string'.
That's because the former only promise the caller it won't modify the reference,
while the later ask for the strong promise (from the caller) that this argument shall never be modified.

In addition, some of those algorithm will either return a portion of the input string or a newly
allocated one.  This may result in a lot of otherwise avoidable allocations.  The runtime decision also
makes it impossible to use in `@nogc` code.
Ultimately, this code could be modified to take a buffer as parameter.
For the time being those functions were turned into templates.