Add an application to test the client websockets implementation against autobahn
Autobahn (http://autobahn.ws/testsuite) is a comprehensive test suite
of the Websocket specifications.
This test won't be run by default as it requires way too much ressources
to put it into the CI, but can trivially be run manually by anyone wanting
to test the websockets implementation for conformity.