Improve usability of DictionaryList with dynamic types
I found it quite cumbersome to work with HTTPServerRequest.context because any value to be assigned to a context key needs to be casted and getting a value required writing something like `req.context.get("foo").get!string`. Using the templated getter, no cast is needed to assign a value and getting it goes like `req.context.get!string("foo")`.