Use requestHTTP for the non-scoped connectWebSocket. Fixes #2169.
This change required a rework of how the connection close and resource management is done to avoid InvalidMemoryOperationError. As a side-effect, this also fixes a leak of TCP connections for connectWebSocket.

The rework in particular makes sure that HTTPClientResponse.disconnect and HTTPServerResponse.finalize is called, and only called from the task that created the web socket. This makes sure that the HTTP client instance is properly returned to the connection pool after the web socket is closed.

Closes #2172.