Remove sanitizeUTF8 call from errorOut
A `string` (or an array of `char`) is supposed to be valid UTF-8,
and potentially-invalid UTF-8 should be an array of `ubyte`.
`errorOut` is only called from 4 places in Vibe.d, all of them in vibe.http.server.
All callers sanitize their input already, so this sould not pose any problems,
unless someone passes invalid `UTF-8` from outside Vibe.d,
but this would blow up just the same if we were to `foreach` over the string.