WebSocket.close: Silently ignore exceptions for close message
When we close our side of the connection, a close message is sent to the
remote side. This can fail if the server cut the connection. In that
case, for example when using OpenSSL as SSL transport, an exception is
thrown. It bubbles up and ends execution of WebSocket.close(), without
resetting the HTTP client/server response objects.

In debug mode, the following warning will appear later:
WARNING: HTTPClientResponse not fully processed before being finalized