Directly construct a valid ProxyStream
A deprecation message used to be triggered in client code because 'opAssign'
was being used, which forwarded to emplace, which forwards to 'std.conv.to',
which attempt to use it.

While the deprecation message has been fixed differently,
there is still value in constructing instances directly
instead of constructing then assigning.