REST: Do not attempt to modify the return value
If the return value of the functioin is const,
trying to re-assiign 'ret' will throw an error inside vibe.d.
Since the only purpose of this separation was for a deprecation message
inside a __traits(compiles), we just use RT.init to give it the correct type,
and merge the two lambdas together.