Change expected TLS test result to match OpenSSL behavior.
The SSL_connect call on the client is successful even if the server rejects the validity of the client certificate (with the client side failing on the first I/O instead). Since we just need to test for one failure here, adjusting the test to expect a client success doesn't do any harm.