Fix file example
Debugging showed that the FileStream is destructed as soon as main is
exited (the task are running later). Interestingly, the bool can be
captured in the closure, but this seems not to work with the refcounted
FileStream.

As discussed on github the best solution is to prolong the lifetime of
the filestream for this example.

Fixes #2547