Move around code to skip over non-existent fields
I've removed the assumption that, when deserializing a struct and
encountering an unknown field, the serializer wants to skip over
a field right now. Instead of deserializeValue(...) directly calling
ignoreValue, it is up to the serializer implementation to do this if the
entry_callback in readDictionary returns false.

This also removes the requirement for serializer to implement a
ignoreValue() function.

This commit also fixes unit tests.