Lessen likelyhood of API breakage reading dictionaries
The entry_callback in readDictionary returns void once again, to undo an
otherwise breaking API change. Instead, the entry_callback itself checks
if the serializer has a skipValue() function and calls that instead.
This might break the API on occasions where serializers have already
defined a skipValue function theirselves, but it will hopefully happen
less often.

Instead on relying on the return value