Update cryptorand.d - add a compile-time check
This pull request depends on https://github.com/dlang/dmd/pull/13169, trying to catch an out of bounds slice at compile time. Currently this would result in a compilation error at line 562, but given the fact that both digestLength!Hash and bufferSize are known at compile time, by replacing if with static if the program will work as expected.