Remove usage of shared static this in examples
It has been discussed and decided in vibe.d/vibe.d#2370 that 'shared static this'
was an early feature that nowadays should go.
The first step to that is to remove it from examples so new users
will not be tempted to use it.

In addition, the examples were tidied up a bit, by always adding a module declaration,
and storing the return values of listen* / runTask whenever applicable.
Some places used spaces instead of tabs, which was also fixed.