openssl: Remove fallback to v1.1.0 version
This fallback is problematic, as there is no sensible default.
A sensible default is something that either is never wrong,
or suits the vast majority of users. This fallback doesn't meet any
of those conditions, hence it is better to be loud and clear about
things not working as expected, and pointing the users in the right direction.
Note that it's unlikely anyone using dub will ever see this message,
as the version generation script already has comprehensive error handling.