CI: Update list of supported compilers
Since the latest release is 2.100, and we support the latest 10 compilers,
we should bump DMD to 2.090. This additionally unblocks the Deimos OpenSSL v3
upgrade, as older DUB versions would try to run the preGenerateCommand in
the directory that contains the dependency, not the dependency itself.

The LDC version was updated to v1.21, despite the latest release being only v1.29,
as v1.30 should be out soon, and DUB packaged with v1.19 is the last one suffering
from the aforementioned bug. Additionally, v1.20 seems to SEGV on some tests.