Fix destructor behavior of BufferedStream.
Since BufferedStream has RAII semantics, the destructor performs an implicit flush to and of the underlying stream, which can in turn throw an exception. In order to still let the destruction happen semantically, any such exception is now rethrown after performing a full clean-up. This means that the data that has not been flushed will be lost, but the state of the object will be so that another destroy() call will be a no-op and no resources are referenced anymore.