Allow sub interface accessors to throw to signal general unavailability.
This enables throwing an exception from a sub interface accessor (e.g. HTTPStatusException with notFound) to control at runtime whether that sub interface is available. Note that a sub interface cannot become available after it has initially thrown an exception, so this is not meant to be a dynamic facility.