chore(README): Remove outdated instructions for OpenSSL
Since 2022 (OpenSSL 3 support in Deimos), the way to select the OpenSSL version
has changed, and those instructions are no longer accurate. The code for it is
now in vibe-stream, which is a separate package.
Instead of updating those instructions, I think it is better to remove them,
as at the time they were added, it was a common problem to have to select
the OpenSSL version. However, with the auto-detect script now in Deimos/OpenSSL,
we are no longer seeing issues about this.