feat(rest): Provide more information on error
I have a use case with an API (Nomad) that doesn't set any Content-Type on error,
returns a 500 (for something that should be 400), and leads to a very user-unfriendly
experience when using the tool. This should provide better out-of-the-box experience
for vibe.web.rest user, by recognizing 'text/plain' messages, and otherwise encoding
the binary data so the caller may be able to do something with it.